requirements.txt
google-generativeai==0.7.2
streamlit==1.36.0
pillow==10.4.0
streamlit-option-menu==0.3.13
